pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "search_docs"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "pandas",
    "openpyxl",
    "pywin32",
    "PyYAML",
    "bteam_utils @ git+https://github.com/bteam-toku/bteam_utils.git",
]

[project.scripts]
search = "search_docs.__main__:main"
search_docs = "search_docs.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }
packages = { find = { where = ["src"] } }
